Move DiscordList.py to discord_ui directory
Renamed and relocated DiscordList.py from shared/ to shared/discord_ui/ for better organization. Updated all imports in affected cogs to reflect the new path.
