+++ bots/derek-bot/cogs/movie_cog.py
@@ -4,7 +4,7 @@ import random
 import logging
 
 from shared.data_manager import DataManager, ListIndexOutOfBounds
-from shared.DiscordList import DiscordList
+from shared.discord_ui.DiscordList import DiscordList
 
 
 class MovieGroupCog(commands.Cog):
